requirements.txt
Django==3.1.6
asgiref==3.3.1
codegen==1.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.2
djoser==2.1.0
drf-nested-routers==0.92.5
pip==21.0.1
pytz==2021.1
setuptools==53.0.0
six==1.15.0
sqlparse==0.4.1
mysqlclient==2.0.3
gunicorn==20.0.4
requests==2.25.1
odfpy==1.4.1